+++ backend/src/test/java/com/pickgo/global/init/TestDataInit.java
@@ -11,10 +11,14 @@ import com.pickgo.domain.member.entity.Member;
 import com.pickgo.domain.member.entity.enums.Authority;
 import com.pickgo.domain.member.entity.enums.SocialProvider;
 import com.pickgo.domain.member.repository.MemberRepository;
-import com.pickgo.domain.performance.entity.*;
+import com.pickgo.domain.performance.entity.Performance;
+import com.pickgo.domain.performance.entity.PerformanceSession;
+import com.pickgo.domain.performance.entity.PerformanceState;
+import com.pickgo.domain.performance.entity.PerformanceType;
 import com.pickgo.domain.performance.repository.PerformanceRepository;
 import com.pickgo.domain.performance.repository.PerformanceSessionRepository;
-import com.pickgo.domain.performance.repository.VenueRepository;
+import com.pickgo.domain.venue.entity.Venue;
+import com.pickgo.domain.venue.repository.VenueRepository;
 import com.pickgo.global.jwt.JwtProvider;
 import com.pickgo.token.TestToken;
 import org.springframework.beans.factory.annotation.Autowired;
